pyproject.toml
[project]
name = "systempy"
version = "0.1.0"
description = "Python application component initialization system"
authors = [{name = "kai3341"}]
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
extra = [
    "celery>=5.5.1",
    "ptpython>=3.0.30",
    "starlette>=0.46.2",
]

[tool.pyright]
include = ["src/"]
reportMissingModuleSource = false
reportArgumentType = false
reportCallIssue = false
reportAssignmentType = false

[tool.ruff.lint]
select=["ALL"]
ignore = ["D", "S101", "TID252", "ANN401", "PTH119"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


[tool.mypy]
python_version = "3.11"
warn_no_return = false

[[tool.mypy.overrides]]
module = [
    'systempy'
]

[dependency-groups]
dev = [
    "mkdocs-material>=9.6.11",
    "mypy>=1.15.0",
    "ruff>=0.11.5",
]
